Migrate validate patterns to utils
